requirements.txt
torch
torchvision
transformers
opencv-python
pillow
PySide6
numpy
matplotlib
python-dotenv